REF: remove need to override get_indexer_non_unique in DatetimeIndexOpsMixin (#33792)
