+++ pandas/core/indexes/datetimelike.py
@@ -8,14 +8,13 @@ import numpy as np
 
 from pandas._libs import NaT, iNaT, join as libjoin, lib
 from pandas._libs.tslibs import timezones
-from pandas._typing import DtypeObj, Label
+from pandas._typing import Label
 from pandas.compat.numpy import function as nv
 from pandas.errors import AbstractMethodError
 from pandas.util._decorators import Appender, cache_readonly, doc
 
 from pandas.core.dtypes.common import (
     ensure_int64,
-    ensure_platform_int,
     is_bool_dtype,
     is_datetime64_any_dtype,
     is_dtype_equal,
@@ -35,7 +34,7 @@ from pandas.core.arrays import DatetimeArray, PeriodArray, TimedeltaArray
 from pandas.core.arrays.datetimelike import DatetimeLikeArrayMixin
 from pandas.core.base import IndexOpsMixin
 import pandas.core.indexes.base as ibase
-from pandas.core.indexes.base import Index, _index_shared_docs, ensure_index
+from pandas.core.indexes.base import Index, _index_shared_docs
 from pandas.core.indexes.extension import (
     ExtensionIndex,
     inherit_names,
@@ -123,12 +122,6 @@ class DatetimeIndexOpsMixin(ExtensionIndex):
     @property
     def is_all_dates(self) -> bool:
         return True
-
-    def _is_comparable_dtype(self, dtype: DtypeObj) -> bool:
-        """
-        Can we compare values of the given dtype to our own?
-        """
-        raise AbstractMethodError(self)
 
     # ------------------------------------------------------------------------
     # Abstract data attributes
@@ -454,21 +447,6 @@ class DatetimeIndexOpsMixin(ExtensionIndex):
 
             # try to find the dates
             return (lhs_mask & rhs_mask).nonzero()[0]
-
-    @Appender(Index.get_indexer_non_unique.__doc__)
-    def get_indexer_non_unique(self, target):
-        target = ensure_index(target)
-        pself, ptarget = self._maybe_promote(target)
-        if pself is not self or ptarget is not target:
-            return pself.get_indexer_non_unique(ptarget)
-
-        if not self._is_comparable_dtype(target.dtype):
-            no_matches = -1 * np.ones(self.shape, dtype=np.intp)
-            return no_matches, no_matches
-
-        tgt_values = target.asi8
-        indexer, missing = self._engine.get_indexer_non_unique(tgt_values)
-        return ensure_platform_int(indexer), missing
 
     # --------------------------------------------------------------------
     # Arithmetic Methods
